move common interface to flint-commons
